+++ CHANGES.rst
@@ -5,10 +5,12 @@ Versions without timestamps mean they are future releases.
 
 2.0.0:
     - drop Pyblosxom support
+    - use ``Attr.default`` instead of ``default_smartypants_attr``
 
 development:
     - add ``--version``
     - add ``install_test`` target for checking package installation
+    - redesign attr input with new ``Attr`` class
 
 1.7.1: 2013-08-14T06:45:59Z
     - fix README-PyPI.rst missing while installing